Swagger added. refactor validations and entities. add @RequiredArgsConstructor annotation. change id -> Long.

+++ src/main/java/com/example/producttestapi/model/CategoryModel.java
@@ -12,7 +12,8 @@ import java.util.List;
 @Setter
 @Getter
 @NoArgsConstructor
-public class CategoryModel {    private int id;
+public class CategoryModel {
+    private Long id;
 
     private String name;
     private List<CategoryModel> categoriesModelList;
