+++ src/main/java/com/example/producttestapi/service/ProductService.java
@@ -7,9 +7,9 @@ import java.util.List;
 
 public interface ProductService {
     List<ProductDto> getAllProducts();
-    ProductDto findProductById(int id);
-    List<ProductDto> getProductsByCategoryID(int categoryID);
+    ProductDto findProductById(Long id);
+    List<ProductDto> getProductsByCategoryID(Long categoryID);
     Product createProduct(Product product);
     Product updateProduct(Product product);
-    void deleteProduct(int id);
+    void deleteProduct(Long id);
 }
